Pathname->Path, Entry as abstract, Magic Engine
Dropped support/pathname, replacing it with the newer native Path and integrating the used features directly into Entry.

Entry is now an abstract class and has an EntryFactory which creates the appropriate subclass. These subclasses are minimally used right now, but they do handle everything needed for nonexistant/notfound files and weird edge case files as well as directories similar to how it worked previously.

The instantiation and calls to the underlying libmagic shard have been moved out of Options and Entry and into a dedicated Engine which wraps the external library to make it presentable.

Regressions:

Nonexistant files provided on the commandline will throw an error. I have a fix for it but this commit is big enough as is.

Entries children size is no longer cached, this could theoretically be an issue with very large directories. I got tired of fighting with Crystal which became convinced that the method which wrapped the value could return nil when I had specifically checked for that. Later I might put it back after I get more important things figured out.

Files and directories are no longer sorted by Entry. I decided that I wanted to have more options for sorting and filtering so it no longer makes sense to do that in Entry. For this commit though it is left unimplemented.