Handle missing files and broken symlinks better
This fixes the issue with nonexistant files on the commandline causing a crash.

This includes another change which stops it from crashing when reading broken symlinks.

----

Actually due to rebasing and partial rewrites, that was fixed in the
previous commit. However, it did provide me the tip to pass false to
info which lets me give that data for broken symlinks.

The new strategy also means no raise/rescue.