Begin rewriting tests, fix resulting bugs.
Tests have been woefully lacking, and I'm starting with the Themer
library. Ideally this will be broken out into its own shard soon.

Discovered a bug wherein if you provided only a single ID to for (rather
than an Array) it would not fallback to default, it would just
raise. This was fixed so now both interfaces to #for have the same
behaviour.