Test correct fallback depth rules.
 - "none" skips settings further down the chain. This is particularly
useful if you want to set something bold at 16-color but use non-bold at
16MM-color. You could also use the "normal" style, but this way has two
important differences: (1) it reduces the number of codes output/parsed
to/by the terminal and (2) it prevents accidentally overwriting a
previous setting, for instance a background color, so you don't have to
keep setting it.

 - Otherwise colors will greedily collect all available settings at the
highest color depth available given the request number of colors.