Replace Color class with Colorset struct.
The goal here is to create an immutable object and reduce the overall
complexity around the storage of color sets.

The API changed slightly, merging the #codes_for method into the #codes
method.

As a consequence #codes are no longer cached.

The intention is to cache the final escape sequences directly in a theme
or static theme cache object instead of the Colorset. In practice, a
set of colors will probably never change after creation, that is why
they were made into structs, therefore this also means that all is really
needed after loading the theme for a given color depth is just a map of
IDs and escape sequences.

I experimented with making Colorcode its own struct instead of a plain
String, but there's no much of a benefit to doing so. I may yet make an
alias for it just to set it apart, but its not particularly important.

Additionally:

The #to_hash method was changed so it doesn't require converting a named
map into a hash.

The #codes method now returns the new_codes_string if it is empty
instead of a new string. I have no idea if that optimization was
something the compiler could have done before and in any case it's
pretty minor, but hey, one less allocation.