Resurrect the corpse of Pathname.
The new Path library is nice, but doesn't actually do most of the things
I need.

However, using macros I greatly reduced the repetitiveness of
everything.

This moved a little code out of Entry but not as much as I was hoping
to.

The directory stuff still should be moved back out. It seemed like a
good idea to get rid of the extra abstraction but then I found a lot of
issues needed to be dealt with both in the Crystal stdlib and with the
unpleasant wayts libc works. Ideally I'd not like to have to do all this
again so it's better to put it all in a lib.